requirements.txt
alpaca-py>=0.20.0
requests>=2.31.0
